Refacfor code
